+++ src/migrations/components/blockCardsSliderTeaser.ts
@@ -1,5 +1,6 @@
 import { ContentfulComponentMigrations, ContentfulMigrationGenerator } from "../types";
 import { migrateBaseBlockFields } from "./block";
+import { getRteValidation, RTE_TYPE_HEADLINE } from "./rte";
 
 const translations = {
     en: {
@@ -29,12 +30,12 @@ const translations = {
 export const getBlockCardsSliderTeaserMigration: ContentfulMigrationGenerator = (
     language
 ): ContentfulComponentMigrations => {
+    const t = translations[language];
+
     return {
         component: "blockCardsSliderTeaser",
         migrations: {
             1: migration => {
-                const t = translations[language];
-
                 const blockCardsSliderTeaser = migration.createContentType(
                     "blockCardsSliderTeaser",
                     {
@@ -72,6 +73,19 @@ export const getBlockCardsSliderTeaserMigration: ContentfulMigrationGenerator = 
                 });
 
                 migrateBaseBlockFields(blockCardsSliderTeaser, language);
+            },
+            2: migration => {
+                const blockCardsSliderTeaser = migration.editContentType("blockCardsSliderTeaser");
+
+                blockCardsSliderTeaser.deleteField("headline");
+
+                blockCardsSliderTeaser.createField("headline", {
+                    type: "RichText",
+                    name: t.blockCardsSliderTeaser.fields.headline,
+                    validations: getRteValidation(RTE_TYPE_HEADLINE),
+                });
+
+                blockCardsSliderTeaser.moveField("headline").afterField("name");
             },
         },
     };