+++ src/migrations/components/blockText.ts
@@ -1,6 +1,6 @@
 import { ContentfulComponentMigrations, ContentfulMigrationGenerator } from "../types";
 import { migrateBaseBlockFields } from "./block";
-import { getRteValidation, RTE_TYPE_STYLED_FONT_AND_LIST } from "./rte";
+import { getRteValidation, RTE_TYPE_HEADLINE, RTE_TYPE_STYLED_FONT_AND_LIST } from "./rte";
 
 const translations = {
     en: {
@@ -28,12 +28,12 @@ const translations = {
 export const getBlockTextMigration: ContentfulMigrationGenerator = (
     language
 ): ContentfulComponentMigrations => {
+    const t = translations[language];
+
     return {
         component: "blockText",
         migrations: {
             1: migration => {
-                const t = translations[language];
-
                 const blockText = migration.createContentType("blockText", {
                     name: t.blockText.name,
                 });
@@ -56,6 +56,19 @@ export const getBlockTextMigration: ContentfulMigrationGenerator = (
                 });
 
                 migrateBaseBlockFields(blockText, language);
+            },
+            2: migration => {
+                const blockText = migration.editContentType("blockText");
+
+                blockText.deleteField("headline");
+
+                blockText.createField("headline", {
+                    type: "RichText",
+                    name: t.blockText.fields.headline,
+                    validations: getRteValidation(RTE_TYPE_HEADLINE),
+                });
+
+                blockText.moveField("headline").afterField("name");
             },
         },
     };